More verbose, check medians
Show more output on the log files by using the correct source() options.
Compare medians from the BigWig generated by Rail-RNA against medians calculated
from the coverage. Since it's hard to do this in Rle world, the median
comparison is restricted to the first 100 thousand base-pairs where the median
by Rail-RNA is non-zero for each of the chromosomes. While the chr with the
largest non-zeros has 6224000 of them, doing the comparison with all non-zero
bases would take much longer.